+++ internal/format/first_pass_visitor.go
@@ -29,27 +29,33 @@ import (
 	"github.com/uber/prototool/internal/text"
 )
 
+var _ proto.Visitor = &firstPassVisitor{}
+
 type firstPassVisitor struct {
 	*baseVisitor
 
-	Syntax                *proto.Syntax
-	Package               *proto.Package
-	Options               []*proto.Option
-	ProbablyCustomOptions []*proto.Option
-	Imports               []*proto.Import
+	Syntax  *proto.Syntax
+	Package *proto.Package
+	Options []*proto.Option
+	Imports []*proto.Import
 
 	haveHitNonComment bool
+
+	updateFileOptions bool
+	goPackageOption   *proto.Option
+	javaPackageOption *proto.Option
 }
 
-func newFirstPassVisitor(config settings.Config) *firstPassVisitor {
-	return &firstPassVisitor{baseVisitor: newBaseVisitor(config.Format.Indent)}
+func newFirstPassVisitor(config settings.Config, updateFileOptions bool) *firstPassVisitor {
+	return &firstPassVisitor{baseVisitor: newBaseVisitor(config.Format.Indent), updateFileOptions: updateFileOptions}
 }
 
 func (v *firstPassVisitor) Do() []*text.Failure {
 	if v.Syntax != nil {
 		v.PComment(v.Syntax.Comment)
 		if v.Syntax.Comment != nil {
-			// special case
+			// special case, we add a newline in between the first comment and syntax
+			// to separate licenses, file descriptions, etc.
 			v.P()
 		}
 		v.PWithInlineComment(v.Syntax.InlineComment, `syntax = "`, v.Syntax.Value, `";`)
@@ -60,9 +66,29 @@ func (v *firstPassVisitor) Do() []*text.Failure {
 		v.PWithInlineComment(v.Package.InlineComment, `package `, v.Package.Name, `;`)
 		v.P()
 	}
-	if len(v.Options) > 0 || len(v.ProbablyCustomOptions) > 0 {
+	if v.updateFileOptions && v.Package != nil {
+		if v.goPackageOption == nil {
+			v.goPackageOption = &proto.Option{Name: "go_package"}
+		}
+		if v.javaPackageOption == nil {
+			v.javaPackageOption = &proto.Option{Name: "java_package"}
+		}
+		v.goPackageOption.Constant = proto.Literal{
+			Source:   packageBasename(v.Package.Name) + "pb",
+			IsString: true,
+		}
+		v.javaPackageOption.Constant = proto.Literal{
+			Source:   "com." + v.Package.Name + ".pb",
+			IsString: true,
+		}
+		v.Options = append(
+			v.Options,
+			v.goPackageOption,
+			v.javaPackageOption,
+		)
+	}
+	if len(v.Options) > 0 {
 		v.POptions(false, v.Options...)
-		v.POptions(false, v.ProbablyCustomOptions...)
 		v.P()
 	}
 	if len(v.Imports) > 0 {
@@ -102,11 +128,20 @@ func (v *firstPassVisitor) VisitOption(element *proto.Option) {
 	// this will only hit file options since we don't do any
 	// visiting of children in this visitor
 	v.haveHitNonComment = true
-	if isProbablyCustomOption(element) {
-		v.ProbablyCustomOptions = append(v.ProbablyCustomOptions, element)
-	} else {
-		v.Options = append(v.Options, element)
+	if v.updateFileOptions {
+		switch element.Name {
+		case "go_package":
+			v.goPackageOption = element
+			return
+		case "java_package":
+			v.javaPackageOption = element
+			return
+		case "java_outer_classname", "java_multiple_files":
+			// ignore
+			return
+		}
 	}
+	v.Options = append(v.Options, element)
 }
 
 func (v *firstPassVisitor) VisitImport(element *proto.Import) {
@@ -127,6 +162,8 @@ func (v *firstPassVisitor) VisitEnum(element *proto.Enum) {
 }
 
 func (v *firstPassVisitor) VisitComment(element *proto.Comment) {
+	// We only print file-level comments before syntax, package, file-level options,
+	// or package if they are at the top of the file
 	if !v.haveHitNonComment {
 		v.PComment(element)
 		v.P()
@@ -180,8 +217,7 @@ func (v *firstPassVisitor) PImports(imports []*proto.Import) {
 	}
 }
 
-func isProbablyCustomOption(option *proto.Option) bool {
-	// you can technically do ie google.protobuf.java_package
-	// but we're not going to handle this as I mean come on
-	return strings.HasPrefix(option.Name, "(")
+func packageBasename(pkg string) string {
+	split := strings.Split(pkg, ".")
+	return split[len(split)-1]
 }