+++ internal/format/transformer.go
@@ -32,7 +32,8 @@ import (
 )
 
 type transformer struct {
-	logger *zap.Logger
+	logger            *zap.Logger
+	updateFileOptions bool
 }
 
 func newTransformer(options ...TransformerOption) *transformer {
@@ -51,15 +52,7 @@ func (t *transformer) Transform(config settings.Config, data []byte) ([]byte, []
 		return nil, nil, err
 	}
 
-	// log statements are at debug level so
-	// this will trigger if debug is set
-	// TODO maybe remove
-	logVisitor := newLogVisitor(t.logger)
-	for _, element := range descriptor.Elements {
-		element.Accept(logVisitor)
-	}
-
-	firstPassVisitor := newFirstPassVisitor(config)
+	firstPassVisitor := newFirstPassVisitor(config, t.updateFileOptions)
 	for _, element := range descriptor.Elements {
 		element.Accept(firstPassVisitor)
 	}
@@ -79,12 +72,12 @@ func (t *transformer) Transform(config settings.Config, data []byte) ([]byte, []
 		}
 	}
 
-	middleVisitor := newMiddleVisitor(config, syntaxVersion == 2)
+	mainVisitor := newMainVisitor(config, syntaxVersion == 2)
 	for _, element := range descriptor.Elements {
-		element.Accept(middleVisitor)
+		element.Accept(mainVisitor)
 	}
-	failures = append(failures, middleVisitor.Do()...)
-	buffer.Write(middleVisitor.Bytes())
+	failures = append(failures, mainVisitor.Do()...)
+	buffer.Write(mainVisitor.Bytes())
 
 	text.SortFailures(failures)