+++ internal/protoc/protoc.go
@@ -80,7 +80,6 @@ func DownloaderWithLogger(logger *zap.Logger) DownloaderOption {
 
 // DownloaderWithCachePath returns a DownloaderOption that uses the given cachePath.
 //
-//
 // The default is ${XDG_CACHE_HOME}/prototool/$(uname -s)/$(uname -m).
 func DownloaderWithCachePath(cachePath string) DownloaderOption {
 	return func(downloader *downloader) {
@@ -107,6 +106,9 @@ type CompileResult struct {
 	// The failures from all calls.
 	Failures []*text.Failure
 	// Will not be set if there are any failures.
+	//
+	// Will only be set if the CompilerWithFileDescriptorSet
+	// option is used.
 	FileDescriptorSets []*descriptor.FileDescriptorSet
 }
 
@@ -118,9 +120,6 @@ type Compiler interface {
 	// and there will be no error. The caller can determine if this is
 	// an error case. If there is any other type of error, or some output
 	// from protoc cannot be interpreted, an error will be returned.
-	//
-	// FileDescriptorSet will only be set if the CompilerWithFileDescriptorSet
-	// option is used.
 	Compile(...*file.ProtoSet) (*CompileResult, error)
 
 	// Return the protoc commands that would be run on Compile.
@@ -142,7 +141,6 @@ func CompilerWithLogger(logger *zap.Logger) CompilerOption {
 }
 
 // CompilerWithCachePath returns a CompilerOption that uses the given cachePath.
-//
 //
 // The default is ${XDG_CACHE_HOME}/prototool/$(uname -s)/$(uname -m).
 func CompilerWithCachePath(cachePath string) CompilerOption {